archive/dasharo_onboarding.md: fix flashrom command
The old command would work if we were updating between two official
Dasharo releases which are properly signed but it is not the case. The
whole bios region needs to be flashed, not just the RW_SECTION_A. The
new command is taken from https://docs.dasharo.com/dasharo-tools-suite/documentation/#features.

Signed-off-by: Filip Gołaś <filip.golas@3mdeb.com>